Building Your First AI Agent with LangChain/Agent Development with LangChain/Agents with Memory/requirements.txt
langchain
langchain-google-genai
google-generativeai
python-dotenv